feat(#474): PlannerDataAccess facade — SQL verplaatst naar 5 repositories
PlannerDataAccess.cs: 1039 → 100 regels (dunne facade)
Alle SQL verplaatst naar FunctionApp/Planner/Repositories/:

PlannerSettingsRepository  (10 methoden):
  GetSpeeltijdAsync, GetVeldenAsync, GetVeldenLookupAsync,
  GetSpeeltijdenLookupAsync, GetTeamLeeftijdLookupAsync,
  GetSunsetAsync, PopulateSunsetTableAsync, GetSeasonEndDateAsync,
  GetLastSyncTimestampAsync, GetVoorkeurTijdenLookupAsync

PlannerAvailabilityRepository (4 methoden):
  GetAvailableFieldsAsync, GetFieldOccupationsAsync,
  GetFieldOccupationsExcludingAsync, GetFieldOccupationsExcludingMatchAsync

PlannerMatchRepository (11 methoden):
  GetTeamMatchesOnDateAsync, GetGeplandeWedstrijdenOnlyAsync,
  FindMatchAsync, FindMatchByOpponentAsync, FindMatchByCodeAsync,
  SavePlannedMatchAsync, SaveHerplanVerzoekAsync,
  MarkeerVervallenGeplandeWedstrijdenAsync, TeamExistsAsync,
  GetFutureMatchesForTeamAsync + DetermineMatchType

TeamRulesRepository (2 methoden):
  GetTeamRulesAsync, GetAllTeamBuffersAsync

AllstarsTestDataRepository (4 methoden):
  GetAllstarsVeldenAsync, GetAllMatchesForDatumAsync,
  UpdateAllstarsMatchAsync, GetTeamleiderContactAsync

Alle 27 tests nog steeds groen. Bestaande callers ongewijzigd.

Versie: 2.13.0.0 → 2.14.0.0 (feat = MINOR bump)
